twitter_trend_view/to_excel_file/requirements.txt
# Function dependencies, for example:
# package>=version

tweepy
openpyxl
pytz
google-cloud
google-cloud-storage

